@@ -4,6 +4,6 @@ import InitMacro
 
 import struct Foundation.Date
 
-@Init public struct Backup {
+@Init public struct Backup: Sendable {
 	public let creationDate: Date
 }
