@@ -3,7 +3,7 @@
 import InitMacro
 
 extension Collection {
-	@Init public struct Access {
+	@Init public struct Access: Sendable {
 		public let level: Level
 		public let isDraggable: Bool
 	}
