@@ -1,7 +1,7 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
 public extension Collection.Access {
-	enum Level {
+	enum Level: Sendable {
 		case readOnly
 		case collaborator(Permissions)
 		case owner
