@@ -2,7 +2,7 @@
 
 import InitMacro
 
-@Init public struct Collaborator {
+@Init public struct Collaborator: Sendable {
 	public let email: String
 	public let fullName: String
 	public let role: Role
