@@ -1,7 +1,7 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
 public extension Collaborator {
-	enum Role {
+	enum Role: Sendable {
 		case owner
 		case member
 		case viewer
