@@ -1,7 +1,7 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
 public extension Collection {
-	enum View {
+	enum View: Sendable {
 		case list
 		case simple
 		case grid
