@@ -4,6 +4,6 @@ import InitMacro
 
 import struct Foundation.URL
 
-@Init public struct Media {
+@Init public struct Media: Sendable {
 	public let url: URL
 }
