@@ -4,7 +4,7 @@ import InitMacro
 
 import struct Foundation.Date
 
-@Init public struct Cache {
+@Init public struct Cache: Sendable {
 	public let status: Status
 	public let size: Int?
 	public let creationDate: Date?
