@@ -1,7 +1,7 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
 public extension Cache.Status {
-	enum Property {
+	enum Property: Sendable {
 		case origin
 		case timeout
 		case size
