@@ -1,7 +1,7 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
 public extension Cache {
-	enum Status {
+	enum Status: Sendable {
 		case ready
 		case retry
 		case failed
