@@ -5,7 +5,7 @@ import InitMacro
 import struct Foundation.URL
 
 extension Raindrop {
-	@Init public struct Info {
+	@Init public struct Info: Sendable {
 		public let title: String
 		public let itemType: ItemType
 		public let excerpt: String?
