@@ -1,6 +1,6 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
-public enum ItemType {
+public enum ItemType: String, Sendable {
 	case link 
 	case article 
 	case image 
