@@ -1,7 +1,7 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
 public extension Raindrop.Sort {
-	enum Criterion {
+	enum Criterion: Sendable {
 		case title
 		case domain
 		case creationDate
