@@ -4,7 +4,7 @@ import InitMacro
 
 import struct Foundation.Date
 
-@Init public struct FileStorage {
+@Init public struct FileStorage: Sendable {
 	public let usedSpace: Int
 	public let totalSpace: Int
 	public let lastCheckpointDate: Date
