@@ -2,7 +2,7 @@
 
 import InitMacro
 
-@Init public struct Group {
+@Init public struct Group: Sendable {
 	public let title: String
 	public let isHidden: Bool
 	public let sortIndex: Int
