@@ -2,7 +2,7 @@
 
 import InitMacro
 
-@Init public struct User {
+@Init public struct User: Sendable {
 	public let fullName: String
 	public let hasProSubscription: Bool
 }
